still making changes

+++ resources/views/livewire/admin/events.blade.php
@@ -1,6 +1,7 @@
 {{-- @livewire('livewire-pagination') --}}
 
 @push('custom-styles')
+<link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@5.2.3/dist/css/bootstrap.min.css">
     <link rel="stylesheet" href="/css/adminmanagement.css">
 @endpush
 
@@ -19,6 +20,8 @@
                     <div class="col-lg-8 col-md-4 col-sm-12 col-xs-12 d-flex justify-content-center align-items-center">
                         <div id="login-nav-bar-links" type="button"
                             onclick=" window.location='{{ route('products.manage') }}'">Products</div>
+
+                            <div id="login-nav-bar-links" type="button" onclick=" window.location='{{ route('categories.manage') }}'">Categories</div>
                         @role('admin')
                             <div id="login-nav-bar-links" type="button"
                                 onclick=" window.location='{{ route('status.manage') }}'">Status</div>
@@ -26,11 +29,27 @@
                                 onclick=" window.location='{{ route('roles.manage') }}'">Roles</div>
                             <div id="login-nav-bar-links" type="button"
                                 onclick=" window.location='{{ route('permissions.manage') }}'">Permissions</div>
-                            <div id="login-nav-bar-links" type="button"
-                                onclick=" window.location='{{ route('users.manage') }}'">Users</div>
+                        @endrole
+
+                        @role('admin')
+                        <div id="login-nav-bar-links" type="button"
+                            onclick=" window.location='{{ route('users.manage') }}'">Users</div>
+                        @else
+                        <div id="login-nav-bar-links" type="button" onclick=" window.location='{{ route('users.manage') }}'">Profile</div>
                         @endrole
                         <div id="login-nav-bar-links" type="button"
                             onclick=" window.location='{{ route('reports.manage') }}'">Reports</div>
+                        <div id="login-nav-bar-links" class="dropdown">
+                            <div class="dropdown-toggle" id="dropdownMenuButton" type="button"
+                            aria-haspopup="true" data-toggle="dropdown" aria-expanded="false">Utilties</div>
+                            <div class="dropdown-menu" aria-labelledby="dropdownMenuButton">
+                                <a class="dropdown-item" onclick=" window.location='{{ route('notices.manage') }}'">Notices</a>
+                                <a class="dropdown-item" onclick=" window.location='{{ route('events.manage') }}'">Events</a>
+                                <a class="dropdown-item" onclick=" window.location='{{ route('faqs.manage') }}'">FAQ's</a>
+                                <a class="dropdown-item" onclick=" window.location='{{ route('suggestions.manage') }}'">Suggestion Box</a>
+                                <a class="dropdown-item" onclick=" window.location='{{ route('slides.manage') }}'">Slides</a>
+                            </div>
+                        </div>
                     </div>
                     <div id="align-right"
                         class="col-lg-2 col-md-4 col-sm-12 col-xs-12 d-flex justify-content-end align-items-end">
@@ -61,11 +80,18 @@
                             </div>
                             <div class="col-sm-4">
 
-                                @role('admin')
+                                @can('create')
                                     <button id="btn-add-new" type="button" class="btn btn-primary" data-toggle="modal"
                                         data-target="#addEventModal"><i class="material-icons">&#xE147;</i> <span>Add New
                                             Event</span></button>
+                                @endcan
+
+                                @role('admin')
+                                <button id="btn-add-new" type="button" class="btn btn-primary" data-toggle="modal" data-target="#addEventModal"><i
+                                        class="material-icons">&#xE147;</i> <span>Add New
+                                        Event</span></button>
                                 @endrole
+                                
                                 <div wire:ignore.self class="modal fade" id="addEventModal" tabindex="-1"
                                     role="dialog" aria-labelledby="addEventModalLabel" aria-hidden="true">
                                     <div class="modal-dialog modal-lg" style="max-width: 80%;" role="document">
@@ -194,6 +220,7 @@
                                         </div>
                                     </div>
                                 </div>
+                               
                             </div>
                         </div>
                     </div>
@@ -230,7 +257,7 @@
                                         </span>
                                     </td>
                                     <td>{{ $upcoming_event->id }}</td>
-                                    <td>{{ $upcoming_event->events_name }}</td>
+                                    <td>{{ $upcoming_event->event_name }}</td>
                                     <td>{{ $upcoming_event->event_description }}</td>
                                     <td>{{ $upcoming_event->venue }}</td>
                                     <td>{{ $upcoming_event->fee }}</td>
@@ -239,18 +266,34 @@
                                     <td>{{ $upcoming_event->start_date }}</td>
                                     <td>{{ $upcoming_event->end_date }}</td>
                                     <td>
-                                        @role('admin')
+                                        @can('update')
                                             <a type="button" href="#updateEventModal" data-toggle="modal"
                                                 data-target="#updateEventModal"
                                                 wire:click="editEvent({{ $upcoming_event->id }})" class="edit"><i
                                                     class="material-icons" data-toggle="tooltip"
                                                     title="Edit">&#xE254;</i></a>
+                                        @endcan
+
+                                        @role('admin')
+                                        <a type="button" href="#updateEventModal" data-toggle="modal" data-target="#updateEventModal"
+                                            wire:click="editEvent({{ $upcoming_event->id }})" class="edit"><i class="material-icons" data-toggle="tooltip"
+                                                title="Edit">&#xE254;</i></a>
+                                        @endrole
+
+                                        @can('delete')
                                             <a type="button" href="#deleteEventModal" data-toggle="modal"
                                                 data-target="#deleteEventModal"
                                                 wire:click="deleteEvent({{ $upcoming_event->id }})" class="delete"
                                                 data-toggle="modal"><i class="material-icons" data-toggle="tooltip"
                                                     title="Delete">&#xE872;</i></a>
+                                        @endcan
+
+                                        @role('admin')
+                                        <a type="button" href="#deleteEventModal" data-toggle="modal" data-target="#deleteEventModal"
+                                            wire:click="deleteEvent({{ $upcoming_event->id }})" class="delete" data-toggle="modal"><i class="material-icons"
+                                                data-toggle="tooltip" title="Delete">&#xE872;</i></a>
                                         @endrole
+                                        
                                     </td>
                                 </tr>
                             @empty
@@ -269,7 +312,7 @@
                 </div>
             </div>
 
-            <!-- Edit Modal HTML -->
+         
             <div wire:ignore.self id="editEventModal" class="modal fade">
                 <div class="modal-dialog modal-dialog-centered" role="document">>
                     <div class="modal-content">
@@ -341,8 +384,9 @@
                     </div>
                 </div>
             </div>
-
-            {{-- Updates the modal --}}
+            
+         
+            
             <div wire:ignore.self id="updateEventModal" class="modal fade">
                 <div class="modal-dialog modal-dialog-centered" role="document">>
                     <div class="modal-content">
@@ -415,9 +459,10 @@
                     </div>
                 </div>
             </div>
-            {{-- end of update modal --}}
+            
 
             <!-- Delete Modal HTML -->
+           
             <div wire:ignore.self id="deleteEventModal" class="modal fade">
                 <div class="modal-dialog modal-dialog-centered" role="document">">
                     <div class="modal-content">
@@ -465,14 +510,11 @@
                     </div>
                 </div>
             </div>
+            
             <!-- end of row -->
-            {{-- </div> --}}
-
         </div>
 
     </body>
-
-
 
 </div>
