+++ resources/views/livewire/admin/suggestion-box.blade.php
@@ -1,6 +1,7 @@
 {{-- @livewire('livewire-pagination') --}}
 
 @push('custom-styles')
+<link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@5.2.3/dist/css/bootstrap.min.css">
     <link rel="stylesheet" href="/css/adminmanagement.css">
 @endpush
 
@@ -28,9 +29,22 @@
                                 onclick=" window.location='{{ route('permissions.manage') }}'">Permissions</div>
                             <div id="login-nav-bar-links" type="button"
                                 onclick=" window.location='{{ route('users.manage') }}'">Users</div>
+                        @else
+                            <div id="login-nav-bar-links" type="button" onclick=" window.location='{{ route('users.manage') }}'">Profile</div>
                         @endrole
                         <div id="login-nav-bar-links" type="button"
                             onclick=" window.location='{{ route('reports.manage') }}'">Reports</div>
+                        <div id="login-nav-bar-links" class="dropdown">
+                            <div class="dropdown-toggle" id="dropdownMenuButton" type="button" aria-haspopup="true" data-toggle="dropdown"
+                                aria-expanded="false">Utilties</div>
+                            <div class="dropdown-menu" aria-labelledby="dropdownMenuButton">
+                                <a class="dropdown-item" onclick=" window.location='{{ route('notices.manage') }}'">Notices</a>
+                                <a class="dropdown-item" onclick=" window.location='{{ route('events.manage') }}'">Events</a>
+                                <a class="dropdown-item" onclick=" window.location='{{ route('faqs.manage') }}'">FAQ's</a>
+                                <a class="dropdown-item" onclick=" window.location='{{ route('suggestions.manage') }}'">Suggestion Box</a>
+                                <a class="dropdown-item" onclick=" window.location='{{ route('slides.manage') }}'">Slides</a>
+                            </div>
+                        </div>
                     </div>
                     <div id="align-right"
                         class="col-lg-2 col-md-4 col-sm-12 col-xs-12 d-flex justify-content-end align-items-end">
@@ -97,15 +111,22 @@
                                     <td>{{ $suggestion->date }}</td>
 
                                     <td>
+
                                         <a type="button" href="#viewModal" data-toggle="modal" data-target="#viewModal"
                                             wire:click.prevent="viewItem({{ $suggestion->id }})"><i
                                                 class="bi bi-eye-fill"></i></a>
-                                        @role('admin')
+                                        @can('delete')
                                             <a type="button" href="#deleteSuggestionModal" data-toggle="modal"
                                                 data-target="#deleteSuggestionModal"
                                                 wire:click="deleteSuggestion({{ $suggestion->id }})" class="delete"
                                                 data-toggle="modal"><i class="material-icons" data-toggle="tooltip"
                                                     title="Delete">&#xE872;</i></a>
+                                        @endcan
+
+                                        @role('admin')
+                                        <a type="button" href="#deleteSuggestionModal" data-toggle="modal" data-target="#deleteSuggestionModal"
+                                            wire:click="deleteSuggestion({{ $suggestion->id }})" class="delete" data-toggle="modal"><i class="material-icons"
+                                                data-toggle="tooltip" title="Delete">&#xE872;</i></a>
                                         @endrole
                                     </td>
                                 </tr>
